Suppress format warnings

+++ genlb.c
@@ -366,16 +366,18 @@ static int genlb_connect(genlb_state_t *s, pid_t tracee) {
   }
 
   if (addrlen < sizeof(struct sockaddr)) {
-    VERBOSE(s, 0, "read: sockaddr: addrlen=%d/%d\n", addrlen,
-            sizeof(struct sockaddr));
+    VERBOSE(s, 0, "read: sockaddr: addrlen=%lu/%lu\n",
+            (long unsigned int)addrlen,
+            (long unsigned int)sizeof(struct sockaddr));
     return -1;
   }
 
   switch (((const struct sockaddr *)&addr)->sa_family) {
   case AF_INET:
     if (addrlen < sizeof(struct sockaddr_in)) {
-      VERBOSE(s, 0, "read: sockaddr_in: addrlen=%d/%d\n", addrlen,
-              sizeof(struct sockaddr_in));
+      VERBOSE(s, 0, "read: sockaddr_in: addrlen=%lu/%lu\n",
+              (long unsigned int)addrlen,
+              (long unsigned int)sizeof(struct sockaddr_in));
       return -1;
     }
 
@@ -439,8 +441,9 @@ static int genlb_connect(genlb_state_t *s, pid_t tracee) {
     char addrstr[INET6_ADDRSTRLEN] = {0};
 
     if (addrlen < sizeof(struct sockaddr_in6)) {
-      VERBOSE(s, 0, "read: sockaddr_in6: addrlen=%d/%d\n", addrlen,
-              sizeof(struct sockaddr_in6));
+      VERBOSE(s, 0, "read: sockaddr_in6: addrlen=%lu/%lu\n",
+              (long unsigned int)addrlen,
+              (long unsigned int)sizeof(struct sockaddr_in6));
       return -1;
     }
